Propagate exception with 'completeExceptionally()'
If there is an error, the CompletableFuture in the Shop class is
not completed normally with 'complete()`, but exceptionally with
'completeExceptionally()` with the raised exception as argument.

This causes the caller of the `get()` method to receive a
java.util.concurrent.ExecutionException holding the original exception
that was raised in the Shop class.