Create CompletableFuture with supplyAsync factory method
Instead of creating the CompletableFuture manually and manually starting
a thread that completes this CompletableFuture, we just pass the code
(as a lambda expression) that is needed to compute the value of the
CompletableFuture to the supplyAsync() factory method.

The CompletableFuture returned by supplyAsync() is equivalent to the
previous CompletableFuture, i.e. it also includes exception handling.