Launching price queries as asynchronous tasks
Each price query is launched as an asynchronous task which returns a
Future<String> immediately and does the actual (blocking) computation in
a separate thread.

The values of all the Futures are then "extracted", which is a blocking
operation.

Since all tasks are *launched* at the same time, the time obtain the
values of *all* tasks equals the computation time of the *longest* task.

(In this case, all the prices are available after 4 seconds, in contrast
to 16 seconds in the case where the prices were queried sequentially.)*

In this version, the blocking "extract" operation is done with the
'get' method, which requires to be surrounded by try-catch. Instead of
'get', 'join' can be used, which does the same thing as 'get' but
doesn't need to be surrounded by try-catch.

* Actually, it turns out that on a machine with 4 cores, only 3
CompletableFuture threads can run in parallel, thus 1-3 shops take 4
seconds, 4-6 shops take 8 seconds, 7-9 shops take 12 seconds, and so on.