Launching price queries as asynchronous tasks
Each price query is launched as an asynchronous task which returns a
Future<String> immediately and does the actual (blocking) computation in
a separate thread.

The values of all the Futures are then "extracted", which is a blocking
operation.

Since all tasks are *launched* at the same time, the time obtain the
values of *all* tasks equals the computation time of the *longest* task.

(In this case, all the prices are available after 4 seconds, in contrast
to 16 seconds in the case where the prices were queried sequentially.)*

In this version, the blocking "extract" operation is done with the
'get' method, which requires to be surrounded by try-catch. Instead of
'get', 'join' can be used, which does the same thing as 'get' but
doesn't need to be surrounded by try-catch.

* Actually, it turns out that the default number of threads equals the
number of CPU cores (4 in our case) whereof 1 thread is used for the
main program, and 3 threads are available for the CompletableFuture
tasks. This means that if there are 4 shops, 1 of these 3 threads has
to execute 2 tasks sequentially, and so the total execution time is
8 seconds, rather than 4 seconds. This problem can be fixed by providing
a custom Executor to 'supplyAsync' with a number of threads >= the
number of shops. This will be implemented in a future commit.