Launching price queries as asynchronous tasks (cont.)
Identical to parent commit, except that 'join' is used to get the values
from the Futures, instead of 'get'.

The 'join' method does not need to be surrounded by try-catch, and is
thus better suited than 'get' to be used in a lambda expression, and it
can also be used as a method reference, as in this version.