Provide custom Executor to supplyAsync
This custom Executor basically is the thread pool from which threads for
the CompletableFuture tasks are drawn.

If no custom executor is set, then a default thread pool is used that
has the same number of threads as there are CPU cores (the value of
Runtime.getRuntime().availableProcessors()). In my case 4.

Of these 4 threads, 1 is used for the main program, and 3 are available
for the CompletableFuture tasks. Thus, if there are 4 shops, one thread
needs to execute two tasks, and the total execution time is 8 seconds,
rather than 2 seconds.

In particular, with 3 threads available for the price query tasks, the
total execution time is 4 seconds for 1-3 shops, 8 seconds for 4-6
shops, 12 seconds for 7-9 shops, etc.

To make the total execution time equal 4 seconds for any number of
shops, there must be one worker thread for each shop, thus 4 threads for
4 shops, 5 threads for 5 shops, etc.

This is achieved by the custom Executor, where we set the number of
threads equal to the number of shops (up to a maximum of 100). Now, each
worker thread has to execute only one price query task, which takes it 4
seconds (the waiting can be done in parallel).

With this solution, the total execution time is 4 seconds for 1-100
shops, 8 seconds for 101-200 shops, etc. (due to the maximum of 100
threads).