Change signatures of getPrice and getPriceAsync
The getPrice and getPriceAsync methods of the Shop class return a String
and Future<String>, respectively, rather than a double and
Future<Double>.

This string has the form "shopName:price:discountCode".

A discount code is detetermined deterministically for each shop/product
combination.

Methods calling getPrice and getPriceAsync in BestPriceFinder, as well
as the tests, have been updated to work with the new return values of
getPrice and getPriceAsync.

The class Discount with the enum Discount.Code has been introduced.