Reorganise tests
The test class is split up into two test classes. One class contains the
tests for querying a single shop, and the other class contains the tests
for querying all shops.

Now, these test classes can be run individually with Gradle in the
following way.

To run only the tests in the BestPriceFinderTestSingleShop class:
  $ gradle test --tests "*SingleShop*"

To run only the tests in the BestPriceFinderTestAllShops class:
  $ gradle test --tests "*AllShops*"

Note that the argument for --tests can be either the full class name or
any pattern matching a test class name, e.g. "*AllShops*".

The command "gradle test" simply runs all the tests that can be found.

Furthermore, the package "org.weibeld.bestprice" has been declared for
all the classes of the project.

Helper methods used by all the test classes have been extracted to
static methods in the TestHelpers class.