Chaining DiscountService query: plain solution
The DiscountService provides a static method applyDiscount, which takes
as argument the price information returned by a shop, and returns a
string containing the discounted price.

The calls to the DiscountService are chained after the calls to the Shops.
In the current implementation of findAllPricesAsync, no CompletionStage
methods are used.

CompletionStage methods are used to schedule actions after the
completion of a CompletableFuture (e.g. launch a new CompletableFuture
with 'thenCompose', or use the value of the CompletableFuture as input
to a Functio with 'thenApply').

In the current implementation, we simply create a list of
CompletableFutures for the price infos, wait for their completation, and
then create a list of CompletableFutures with the discount info, wait
for their completion, and return their values.