Chaining DiscountService: sophisticated solution
The implementation of 'findAllPricesAsync' produces the same result as
in the parent commit, but it is made more concise by using the
CompletionStage methods 'thenApply' and 'thenCompose'.

First, 'thenApply' is used to transform a CompletableFuture<String>
(representing the price info from the Shop) to a CompletableFuture<Quote>.

Then, 'thenCompose' is used to transform this CompletableFuture<Quote>
to a CompletableFuture<String> by supplying it as argument a Function
which returns a CompletableFuture<String>. Once the value of the
CompletableFuture<Quote> is available, this function is called with the
Quote value as argument.

The body of this function consists of the CompletableFuture.supplyAsync
method, which in turn takes a Supplier as argument.

In the end, we create a List of  CompletableFuture<String>, where each
string represents the discounted price information from the
DiscountService. Finally we wait for the completion of these
CompletableFutures and return their values as a list of strings.