Print results as soon as they appear
The 'findAllPricesAsync' method does not return a list of strings
anymore, but a stream of CompletableFuture<String>, where each string
represents a discounted price as returned by the DiscountService.

The invoking code can then use this stream of CompletableFuture<String>
as it pleases.

In our case we add a consumer action with 'thenAccept' to each
CompletableFuture<String> which returns a CompletableFuture<Void>. The
action is executed once the CompletableFuture<String> completes and
it consists of printing the resulting string.

Finally, all these CompletableFuture<Void> are collected in an array and
supplied to the 'allOf' method which returns a CompletableFuture<Void>
which completes once all of its CompletableFuture have completed. This
provides a way to wait for the completion of all the previous
CompletableFuture<Void>.