Add currency conversion from USD to EUR
All the shops return their prices in USD, however, our BestPriceFinder
wants to return them in EUR.

Therefore, after applying the discount, the BestPriceFinder makes
another (independent) asynchronous call to the ExchangeService for
retrieving the current USD to EUR exchange rate, and then combines the
result of the DiscountService call with the result of the
ExchangeService call, i.e. multiplying the price in USD with the USD to
EUR exchange rate.

This combination is done by calling 'thenCombine' on the CompletableFuture
holding the result of the DiscountService call.

The return value of Quote.parse and of the DiscountService has been
changed from a String to a ShopPrice object, consisting of shop name,
price, and currency. This is to make the conversion of a price to
another currency more straighforward.