De-sequentialise exchange rate query
In the parent commit, the query for the USDEUR exchange rate is executed
AFTER both, the price query and the discount query. So, if each query
takes 2 seconds, than the total execution time is 6 seconds. This is
however not necessary, because the exchange rate query does not depend
on any of the price query or discount query, so it can be executed in
parallel with them.

The sequential execution of the exchange rate query was because it was
launched in 'thenCombine' of the pipeline. Now it has been moved as an
independent query to the beginning of the method, and the total
execution time is now 4 seconds, if each of the three queries takes 2
seconds.